Fix revision question being treated as the answer
Somehow, the Markdown parser is rendering this question as if it were a
list item, and thus as the correct answer to the revision question,
instead of as the question itself.  Hopefully, changing this line to an
`<h4>` will prevent the misinterpretation. This will clarify the
question for the user and cause the grader to see the *intended* answer
as the correct one.