Update use of ntpd
The revision question was using `ntpdate` which is depricated:
```
Disclaimer: The functionality of this program is now available in the ntpd program. See the -q command line option in the ntpd - Network Time Protocol (NTP) daemon page. After a suitable period of mourning, the ntpdate program is to be retired from
```
[As seen in](https://linux.die.net/man/8/ntpdate)