Update go-for-loops.md with disclaimers and data summaries
Added disclaimers and data processing summaries to the go-for-loops documentation.Disclaimers
THE SITE IS PROVIDED BY PACE "AS IS" AND WITH ALL FAULTS. PACE MAKES NO REPRESENTATIONS OR WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED WITH REGARD TO THE SITE, ITS CONTENTS, OR ITS OPERATION. TO THE FULLEST EXTENT PERMITTED BY LAW, PACE DISCLAIMS ALL WARRANTIES, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, OR NON-INFRINGEMENT.

 

WITHOUT LIMITING THE GENERALITY OF THE FOREGOING, PACE SPECIFICALLY STATES THAT THE USE OF THE SITE IS AT YOUR OWN RISK. PACE DISCLAIMS ANY WARRANTY WITH REGARD TO THE OPERATION OF THE SITE, THE ACCURACY OR TIMELINESS OF THE SITE, OR ITS CONTENTS OR THE INFORMATION OR MATERIALS CONTAINED THEREIN, ANY WARRANTY THAT THE SITE OR ITS CONTENTS OR THE INFORMATION OR MATERIALS CONTAINED THEREIN WILL BE FREE OF ERRORS, VIRUSES OR OTHER COMPONENTS THAT MAY INFECT, HARM, OR CAUSE DAMAGE TO YOUR COMPUTER EQUIPMENT OR ANY OTHER PROPERTY, OR ANY WARRANTY THAT YOUR USE OF THE SITE OR ITS CONTENT OR THE INFORMATION AND MATERIALS CONTAINED HEREIN WILL BE UNINTERRUPTED.

 

WITHOUT LIMITING THE GENERALITY OF THE FOREGOING, PACE SPECIFICALLY INFORMS YOU THAT PACE MAKES NO WARRANTIES, REPRESENTATIONS, OR GUARANTEES AS TO THE ACCURACY, TRUTHFULNESS, OR RELIABILITY OF ANY POSTED INFORMATION.




Exclusion of damages and limitation of liability
UNDER NO CIRCUMSTANCES SHALL PACE, ITS SUPPLIERS, VENDORS OR LICENSORS BE LIABLE FOR ANY DIRECT, SPECIAL, INCIDENTAL, INDIRECT, ECONOMIC OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, DAMAGES FOR LOSS OF DATA, LOSS OF USE OR LOST PROFITS) ARISING OUT OF OR CONNECTED WITH THE SITE (INCLUDING BUT NOT LIMITED TO ITS OPERATION, ITS CONTENTS OR THE INFORMATION OR MATERIALS CONTAINED THEREIN, OR THE USE OR INABILITY TO USE THE SITE, OR THE USE OR INABILITY TO USE ANY OTHER SITE LINKED TO THE SITE, OR ANY CONTENT CONTAINED IN ANY SUCH SITE, OR THESE TERMS OF USE, OR ANY OTHER POLICY APPLICABLE TO THE SITE) EVEN IF PACE, THEIR SUPPLIERS, OR LICENSORS HAVE BEEN NOTIFIED OF THE POSSIBILITY OF ANY DAMAGES. DUE TO THE FACT THAT SOME STATES DO NOT ALLOW THE EXCLUSION OR LIMITATION OF LIABILITY FOR CONSEQUENTIAL OR INCIDENTAL DAMAGES, THE ABOVE LIMITATION MAY NOT APPLY TO YOU. IN SUCH STATES, THE RESPECTIVE LIABILITY OF PACE, ITS SUPPLIERS, VENDORS, AND LICENSORS SHALL BE LIMITED TO THE GREATEST EXTENT PERMITTED BY APPLICABLE LAW.







Data Collection and Processing Summary
When you create an iLok User ID, or one is created for you by a 3rd party, or you use iLok License Manager or any of our license management services to view or manage your licenses in any way:

Purpose: To provide you access and control over software licenses granted to you by 3rd Party Software Publishers and Distributors.

Processing made:

Creation and administration of your iLok User ID (including to maintain correct and updated data). Note, in some cases it is possible that a 3rd party creates your account on your behalf. You may always claim your account at https://www.ilok.com/#!provisional
Confirmation of your contact information, specifically verification of your email address
Administration of your licenses connected to your iLok User ID in order to allow you to manage your licenses and activate them so that you may use your software
Sending you or the publishers information related to any status of the licenses in order to provide you support
Personal data processed:

E-mail address
iLok User ID and password (never shared with anyone)
Birthday day and month and Security Question and Answer for account use (never shared with anyone)
License information and status of licenses granted to you by 3rd party software publishers or distributors
Legal basis:

The processing is necessary in order to create and administer your account in accordance with our Terms & Conditions and for us to thereby fulfil our contract with publishers so that you may use their licenses. If you do not provide us information of your e-mail address, your user name and password, and items above, we will not be able to create your account
To administer information about your license management account so that publishers may support you and give you updates and their legitimate needs to have access to such information to provide such support
When you visit or shop on our Website or use iLok License Manager or any of our license management services:

Purpose: To administer and complete your purchase (including orders for updates and upgrades of previously purchased products)

Processing made:

Creation and administration of your iLok User ID (including to maintain correct and updated data)
Confirmation of your contact information
Administration of your purchase history connected to your iLok User ID in order to provide you with the updates, upgrades and discounts you are eligible for
Sending you information about your licenses and status in case of a support incident or any issue that requires informing you of changes regarding your licenses
Personal data processed:

E-mail address
iLok User ID and password
Purchase history (for example, what products and updates you have purchased from ilok.com directly)
Legal basis:

The processing is necessary in order to create and administer your account in accordance with our Terms & Conditions and for us to thereby fulfil our contract with 3rd Party software publishers. If you do not provide us information of your e-mail address, your user name and password and security question/answer, we will not be able to create your account
To administer information about your license management account and your purchase history as well as sending you information about updates, upgrades and new releases of your purchased products is necessary for the purposes of our and your legitimate interest of having updated products
When We Communicate with Each Other and Publishers/Distributors of the software licenses you use.

Purpose: To provide you with support and to communicate with you in relation to such and similar matters. In addition, we supply publishers of software you use with information they need to provide you with customer service concerning your account and the licenses of theirs that you have been granted.

Processing made:

Confirmation of your identity and contact information
Communication and correspondence with you by e-mail, or our digital channels
Reply to your comments and questions, for example in a support incident
Internal administration of your support matter or comments or questions to our customer service. Note in some cases this may require confirmation with the software publisher related to the license in question. Such support may be initiated by you and come to us either directly from you or a 3rd party publisher (the owner of the licensed software and IP)
Personal data processed:

Contact information – name, e-mail address, etc.
Information that you provide us for example information about your iLok.com account, your products or the questions you ask about our products and services
If you contact us in our digital channels (Facebook, Instagram, Twitter, YouTube): your profile name and photo if so enabled by you on such platforms
Legal basis:

The processing is necessary for the purposes of our legitimate interest of helping you with your support matter or reply to your comments and questions and in order to meet your request for communication with us